Make script to produce course list json from Courses.xml
The script is far from perfect.  The json that it produces has an extra comma on the second to last line which must be removed in order for the json to be valid.

Also, I wanted to include the course name, but having such long option text doesn't seem to work.